backend/requirements.txt
django
djangorestframework
pandas
scikit-learn
joblib